[LIB-154] Change coordinates to float

+++ MNE/fiff/fiff_ch_pos.h
@@ -109,10 +109,10 @@ public:
 
 public:
     fiff_int_t   coil_type; /**< The kind of the coil. */
-    Eigen::Vector3d r0;     /**< Coil coordinate system origin */
-    Eigen::Vector3d ex;     /**< Coil coordinate system x-axis unit vector */
-    Eigen::Vector3d ey;     /**< Coil coordinate system y-axis unit vector */
-    Eigen::Vector3d ez;     /**< Coil coordinate system z-axis unit vector */
+    Eigen::Vector3f r0;     /**< Coil coordinate system origin */
+    Eigen::Vector3f ex;     /**< Coil coordinate system x-axis unit vector */
+    Eigen::Vector3f ey;     /**< Coil coordinate system y-axis unit vector */
+    Eigen::Vector3f ez;     /**< Coil coordinate system z-axis unit vector */
 
 // ### OLD STRUCT ###
 // /** Measurement channel position and coil type. *
